requirements.txt
absl-py==2.3.1
appnope==0.1.4
asttokens==3.0.1
astunparse==1.6.3
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
dm-tree==0.1.9
docstring_parser==0.17.0
einops==0.8.1
etils==1.13.0
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.1.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
immutabledict==4.2.2
importlib_resources==6.5.2
ipykernel==7.1.0
ipython==8.38.0
jedi==0.19.2
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupytext==1.18.1
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.1.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
promise==2.3
prompt_toolkit==3.0.52
protobuf==4.25.8
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
simple-parsing==0.1.7
six==1.17.0
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-datasets==4.9.9
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metadata==1.17.3
tensorflow-metal==1.2.0
termcolor==3.3.0
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.4.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.2.14
Werkzeug==3.1.5
wrapt==2.0.1
zipp==3.23.0
